+++ CHANGELOG.md
@@ -1,8 +1,10 @@
+## Unreleased
+
+- Queue time capture for Rack ([#2838](https://github.com/getsentry/sentry-ruby/pull/2838))
+
 ## 6.3.0
 
 ### Features
-
-- Queue time capture for Rack ([#2838](https://github.com/getsentry/sentry-ruby/pull/2838))
 
 - Implement new `Sentry.metrics` functionality ([#2818](https://github.com/getsentry/sentry-ruby/pull/2818))